+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.15.0
 
+### Enhancements
+- Queries can now be constructed by chosing concepts in another data source that are linked to the original query data source
+
 ## 0.14.2 (June 19, 2013)
 
 ### Enhancements